refactor: Change for util module use
