+++ util/src/main/java/com/apighost/util/file/TimeUtils.java
@@ -1,4 +1,4 @@
-package com.apighost.util;
+package com.apighost.util.file;
 
 import java.time.Instant;
 import java.time.LocalDateTime;
